chore: rename test files to match the move of test case definiton
Separate commit to retain git history.
