+++ src/crap/__tests__/cjs/class.ts
