+++ src/crap/__tests__/cjs/index.ts
@@ -1,8 +1,8 @@
 import { CrapReportTest } from "../crap-report.test.js";
-import { classTestCases } from "./class.test.js";
-import { ifElseTestCases } from "./if-else.test.js";
-import { nestedFunctionsTestCases } from "./nested-functions.test.js";
-import { reExportTestCases } from "./re-export.test.js";
+import { classTestCases } from "./class.js";
+import { ifElseTestCases } from "./if-else.js";
+import { nestedFunctionsTestCases } from "./nested-functions.js";
+import { reExportTestCases } from "./re-export.js";
 
 export const cjsTestCases: CrapReportTest[] = [
     ...classTestCases,
