+++ src/crap/__tests__/cjs/nested-functions.ts
