+++ src/crap/__tests__/esm/class-expression.ts
