+++ src/crap/__tests__/esm/multi-function-line.ts
