+++ src/crap/__tests__/esm/multi-line-function-definition.ts
